webscraper/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# webscraper.py: 6
google_api_python_client == 1.7.8

# webscraper.py: 8
google_auth == 1.6.2

# webscraper.py: 7
google_auth_oauthlib == 0.2.0

# webscraper.py: 9
selenium == 3.141.0

# webscraper.py: 10
webdriver_manager == 1.7
